refactor: 디렉터리 구조 변경
- interfaces 디렉터리 추가
- .interface 파일 위치를 interfaces 디렉터리로 변경
- .interface 변경으로 인한 import문 수정
