+++ src/index.ts
@@ -1,3 +1,3 @@
+export * from './interfaces';
 export * from './message-broker';
 export * from './types';
-export * from './broker.interface';
