+++ src/interfaces/subscription.interface.ts
@@ -1,0 +1,6 @@
+import { MessageHandler } from '../types';
+
+export interface Subscription<TMessage> {
+    readonly id: string;
+    readonly handler: MessageHandler<TMessage>;
+}
